LayoutTests/platform/chromium/http/tests/security/cross-origin-local-storage-wk1-expected.txt

CONSOLE MESSAGE: Invalid name for preference: WebKitStorageBlockingPolicy
CONSOLE MESSAGE: Invalid name for preference: WebKitStorageBlockingPolicy
This iframe should return a security error:


This iframe should not return any errors:



--------
Frame: '<!--framePath //<!--frame0-->-->'
--------
No exception

--------
Frame: '<!--framePath //<!--frame1-->-->'
--------
No exception
